gr-digital: improved performance and capability to 63-bit registers. Now uses __builtin_parityl to determine next bit.

+++ gr-digital/lib/descrambler_bb_impl.h
@@ -35,7 +35,7 @@ private:
     digital::lfsr d_lfsr;
 
 public:
-    descrambler_bb_impl(int mask, int seed, int len);
+    descrambler_bb_impl(uint64_t mask, uint64_t seed, uint8_t len);
     ~descrambler_bb_impl();
 
     int work(int noutput_items,
